Working on SQL part + test tool
The SQL part has some model problems.
One of them is the mdtSqlWindow, mdtSqlFor, etc...
For the moment, a cleaner API is started, in the new mdtdatabase package.
Old classes are now suffixed Old, and will be removed as soon as tools (cable list + test tool) are migrated to new API.