Working on database module (SQL table schema)
Now, mdtSqlSchemTable is again able to generate SQL statements to drop and create tables.
Index and foreign key syntax is cleaner, and indexes are now really created.
When creating a foreign key, a flag can be used for index creation. A index (with a generated name) will also be created automatically.