Prevent data race when logging errors to the console
In lib Error, the Logger did run all its backends in a separate thread.
The ConsoleBackend uses qDebug(), and nothing tells us that it is
thread safe in the Qt documentation.
One case appeard in build number 70 in Travis CI.
Now, Logger can add backends to the main thread, and uses signal/slots,
so that errors can be logged from any thread.
Adding backends to the logger's separate thread works as before.