Remove installing Debian copyright
To generate conform Debian packages,
one of the requirement is to provide a specific copyright file.
Until now, a install(FILE) was used for this.
The side effect is: when install Mdt a other way than generating
Debian packages, a set of copies of the file have been installed.
Note: the generated Debian packages have never be conform at all.